First version with semi-working permission checks
Now calls to open() will walk the policy struct and confirm the client
process has read permissions for the given object.

Many things are still missing, including recursive object definitions
and actual permissions checks (not just read access).

There is also a bug as the method of checking for usernames is not
reenterant.